+++ ReleaseNotes/release_notes.md
@@ -5,6 +5,7 @@
    1. make use of `get_env` and `has_env` functions alsmost everywhere insted of `os.environ`
    1. small reshuffle of code between `ostap.core.core` and `ostap.utils.basic`
    1. first step towards usage of `ipyparallel` for parallel processing: ad trivial test `test_parallel_ipyparallel.py`
+   1. make use of `ipyparallel` parallelisation
    
 ## Backward incompatible: